Refactor Loritta's config, wire the new codebase to the old codebase
Now it works

+++ discord/loritta-bot-common/src/main/kotlin/net/perfectdreams/loritta/morenitta/utils/ClusterOfflineException.kt
@@ -1,3 +1,3 @@
 package net.perfectdreams.loritta.morenitta.utils
 
-data class ClusterOfflineException(val id: Long, val name: String) : RuntimeException("Cluster $id ($name) is offline")
+data class ClusterOfflineException(val id: Int, val name: String) : RuntimeException("Cluster $id ($name) is offline")
